port: linux support; horrible hack to avoid segmented address funny moments
linux support only tested a little bit in a VM

+++ include/PR/os_libc.h
@@ -80,9 +80,9 @@ extern "C" {
 
 /* byte string operations */
 
-extern void     bcopy(const void *, void *, int);
-extern int      bcmp(const void *, const void *, int);
-extern void     bzero(void *, int);
+extern void     bcopy(const void *, void *, size_t);
+extern int      bcmp(const void *, const void *, size_t);
+extern void     bzero(void *, size_t);
 
 /* Printf */
